Restructure and improve tests, remove duplication
