+++ src/test/java/de/tum/in/test/api/HiddenPublicTest.java
@@ -1,49 +1,46 @@
 package de.tum.in.test.api;
 
-import static de.tum.in.testutil.CustomConditions.*;
-import static org.junit.platform.engine.discovery.DiscoverySelectors.selectClass;
+import static de.tum.in.test.testutilities.CustomConditions.*;
 import static org.junit.platform.testkit.engine.EventConditions.*;
 
 import java.lang.annotation.AnnotationFormatError;
 
-import org.junit.jupiter.api.BeforeAll;
-import org.junit.jupiter.api.Tag;
-import org.junit.platform.testkit.engine.EngineTestKit;
 import org.junit.platform.testkit.engine.Events;
 import org.opentest4j.AssertionFailedError;
 
+import de.tum.in.test.testutilities.TestTest;
+import de.tum.in.test.testutilities.UserBased;
+import de.tum.in.test.testutilities.UserTestResults;
 import de.tum.in.testuser.HiddenPublicUser;
-import de.tum.in.testutil.TestTest;
 
+@UserBased(HiddenPublicUser.class)
 class HiddenPublicTest {
-	private final String testHiddenCustomDeadlineFuture = "testHiddenCustomDeadlineFuture";
-	private final String testHiddenCustomDeadlinePast = "testHiddenCustomDeadlinePast";
-	private final String testHiddenNormal = "testHiddenNormal";
-	private final String testHiddenIncomplete = "testHiddenIncomplete";
-	private final String testPublicCustomDeadline = "testPublicCustomDeadline";
-	private final String testPublicNormal = "testPublicNormal";
-	private final String testPublicIncomplete = "testPublicIncomplete";
 
+	@UserTestResults
 	private static Events tests;
 
-	@BeforeAll
-	@Tag("test-test")
-	static void verifyDeadline() {
-		var results = EngineTestKit.engine("junit-jupiter").selectors(selectClass(HiddenPublicUser.class)).execute();
-		tests = results.testEvents();
-
-		results.containerEvents().assertStatistics(stats -> stats.started(2).succeeded(2));
-		tests.assertStatistics(stats -> stats.started(5));
-	}
+	private final String testHiddenCustomDeadlineFuture = "testHiddenCustomDeadlineFuture";
+	private final String testHiddenCustomDeadlinePast = "testHiddenCustomDeadlinePast";
+	private final String testHiddenIncomplete = "testHiddenIncomplete";
+	private final String testHiddenNormal = "testHiddenNormal";
+	private final String testPublicCustomDeadline = "testPublicCustomDeadline";
+	private final String testPublicIncomplete = "testPublicIncomplete";
+	private final String testPublicNormal = "testPublicNormal";
 
 	@TestTest
-	void test_testPublicNormal() {
-		tests.assertThatEvents().haveExactly(1, event(test(testPublicNormal), finishedSuccessfullyRep()));
+	void test_testHiddenCustomDeadlineFuture() {
+		tests.assertThatEvents().haveExactly(1,
+				testFailedWith(testHiddenCustomDeadlineFuture, AssertionFailedError.class));
 	}
 
 	@TestTest
 	void test_testHiddenCustomDeadlinePast() {
 		tests.assertThatEvents().haveExactly(1, event(test(testHiddenCustomDeadlinePast), finishedSuccessfullyRep()));
+	}
+
+	@TestTest
+	void test_testHiddenIncomplete() {
+		tests.assertThatEvents().doNotHave(event(test(testHiddenIncomplete)));
 	}
 
 	@TestTest
@@ -57,19 +54,12 @@ class HiddenPublicTest {
 	}
 
 	@TestTest
-	void test_testHiddenCustomDeadlineFuture() {
-		tests.assertThatEvents().haveExactly(1,
-				testFailedWith(testHiddenCustomDeadlineFuture, AssertionFailedError.class));
-	}
-
-	@TestTest
-	void test_testHiddenIncomplete() {
-		tests.assertThatEvents().doNotHave(event(test(testHiddenIncomplete)));
-	}
-
-	@TestTest
 	void test_testPublicIncomplete() {
 		tests.assertThatEvents().doNotHave(event(test(testPublicIncomplete)));
 	}
 
+	@TestTest
+	void test_testPublicNormal() {
+		tests.assertThatEvents().haveExactly(1, event(test(testPublicNormal), finishedSuccessfullyRep()));
+	}
 }
