+++ src/test/java/de/tum/in/testuser/DeadlineAdditionsUser.java
@@ -9,55 +9,6 @@ import de.tum.in.test.api.jupiter.PublicTest;
 @Deadline("2000-01-01 00:00")
 public class DeadlineAdditionsUser {
 
-	@PublicTest
-	@ActivateHiddenBefore("2200-01-01 00:00")
-	void testHiddenActive() {
-		// nothing
-	}
-
-	@HiddenTest
-	@ExtendedDeadline("50000d")
-	void testHiddenExtendedNormal() {
-		// nothing
-	}
-
-	@HiddenTest
-	@ActivateHiddenBefore("2200-01-01 00:00")
-	@ExtendedDeadline("50000d")
-	void testHiddenExtendedActive() {
-		// nothing
-	}
-
-	@HiddenTest
-	@ActivateHiddenBefore("2000-01-01 00:00")
-	@ExtendedDeadline("50000d")
-	void testHiddenExtendedInactive() {
-		// nothing
-	}
-
-	@PublicTest
-	@ActivateHiddenBefore("2000-01-01 00:00")
-	void testHiddenInactive() {
-		// nothing
-	}
-
-	@PublicTest
-	void testHiddenNormal() {
-		// nothing
-	}
-
-	@PublicTest
-	@ActivateHiddenBefore("2200-01-01 00:00")
-	void testPublicActive() {
-		// nothing
-	}
-
-	@PublicTest
-	@ExtendedDeadline("50000d")
-	void testPublicExtended() {
-		// nothing
-	}
-
 	@HiddenTest
 	@ActivateHiddenBefore("2200-01-01 00:00")
 	@Deadline("2200-01-01 16:00")
@@ -115,6 +66,55 @@ public class DeadlineAdditionsUser {
 	@ExtendedDeadline("50000d")
 	@Deadline("2000-01-01 16:00")
 	void testHidden_CustomDeadlinePastExtendedNormal() {
+		// nothing
+	}
+
+	@PublicTest
+	@ActivateHiddenBefore("2200-01-01 00:00")
+	void testHiddenActive() {
+		// nothing
+	}
+
+	@HiddenTest
+	@ActivateHiddenBefore("2200-01-01 00:00")
+	@ExtendedDeadline("50000d")
+	void testHiddenExtendedActive() {
+		// nothing
+	}
+
+	@HiddenTest
+	@ActivateHiddenBefore("2000-01-01 00:00")
+	@ExtendedDeadline("50000d")
+	void testHiddenExtendedInactive() {
+		// nothing
+	}
+
+	@HiddenTest
+	@ExtendedDeadline("50000d")
+	void testHiddenExtendedNormal() {
+		// nothing
+	}
+
+	@PublicTest
+	@ActivateHiddenBefore("2000-01-01 00:00")
+	void testHiddenInactive() {
+		// nothing
+	}
+
+	@PublicTest
+	void testHiddenNormal() {
+		// nothing
+	}
+
+	@PublicTest
+	@ActivateHiddenBefore("2200-01-01 00:00")
+	void testPublicActive() {
+		// nothing
+	}
+
+	@PublicTest
+	@ExtendedDeadline("50000d")
+	void testPublicExtended() {
 		// nothing
 	}
 }
