+++ src/test/java/de/tum/in/testuser/PackageAccessUser.java
@@ -1,0 +1,59 @@
+package de.tum.in.testuser;
+
+import java.util.concurrent.TimeUnit;
+
+import org.junit.jupiter.api.MethodOrderer.Alphanumeric;
+import org.junit.jupiter.api.TestMethodOrder;
+
+import de.tum.in.test.api.BlacklistPackage;
+import de.tum.in.test.api.BlacklistPath;
+import de.tum.in.test.api.MirrorOutput;
+import de.tum.in.test.api.MirrorOutput.MirrorOutputPolicy;
+import de.tum.in.test.api.PathType;
+import de.tum.in.test.api.StrictTimeout;
+import de.tum.in.test.api.WhitelistPackage;
+import de.tum.in.test.api.WhitelistPath;
+import de.tum.in.test.api.jupiter.PublicTest;
+import de.tum.in.test.api.localization.UseLocale;
+import de.tum.in.testuser.subject.PackageAccessPenguin;
+
+@MirrorOutput(MirrorOutputPolicy.DISABLED)
+@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
+@TestMethodOrder(Alphanumeric.class)
+@UseLocale("en")
+@WhitelistPath(value = "target/**", type = PathType.GLOB)
+@BlacklistPath(value = "**Test*.{java,class}", type = PathType.GLOB)
+@SuppressWarnings("static-method")
+public class PackageAccessUser {
+
+	@BlacklistPackage("java.util")
+	@PublicTest
+	public void package_aBlacklistingUtil() {
+		PackageAccessPenguin.useArrayList();
+	}
+
+	@BlacklistPackage("java.**")
+	@PublicTest
+	public void package_bBlacklistingJava() {
+		PackageAccessPenguin.useArrayList();
+	}
+
+	@BlacklistPackage("**")
+	@PublicTest
+	public void package_cBlacklistingAll() {
+		PackageAccessPenguin.useArrayList();
+	}
+
+	@BlacklistPackage("**")
+	@WhitelistPackage("java.util")
+	@PublicTest
+	public void package_dBlackAndWhitelisting() {
+		PackageAccessPenguin.useArrayList();
+	}
+
+	@BlacklistPackage("java.util")
+	@PublicTest
+	public void package_eBlackPenguinAgain() {
+		PackageAccessPenguin.useArrayList();
+	}
+}
