+++ src/test/java/de/tum/in/testuser/StrictTimeoutUser.java
@@ -14,38 +14,6 @@ import de.tum.in.test.api.StrictTimeout;
 public class StrictTimeoutUser {
 
 	@Test
-	@StrictTimeout(1)
-	void testOneSecondFail() throws InterruptedException {
-		Thread.sleep(10_000);
-	}
-
-	@Test
-	@StrictTimeout(1)
-	void testOneSecondSuccess() throws InterruptedException {
-		Thread.sleep(800);
-	}
-
-	@Test
-	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
-	void testMethodFailNormal() throws InterruptedException {
-		Thread.sleep(500);
-	}
-
-	@Test
-	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
-	void testMethodFailLoop() {
-		int x = 0;
-		while (true)
-			x++;
-	}
-
-	@Test
-	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
-	void testMethodSuccess() throws InterruptedException {
-		Thread.sleep(200);
-	}
-
-	@Test
 	void testClassFailLoop() {
 		int x = 0;
 		while (true)
@@ -60,5 +28,37 @@ public class StrictTimeoutUser {
 	@Test
 	void testClassSuccess() throws InterruptedException {
 		Thread.sleep(20);
+	}
+
+	@Test
+	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
+	void testMethodFailLoop() {
+		int x = 0;
+		while (true)
+			x++;
+	}
+
+	@Test
+	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
+	void testMethodFailNormal() throws InterruptedException {
+		Thread.sleep(500);
+	}
+
+	@Test
+	@StrictTimeout(value = 300, unit = TimeUnit.MILLISECONDS)
+	void testMethodSuccess() throws InterruptedException {
+		Thread.sleep(200);
+	}
+
+	@Test
+	@StrictTimeout(1)
+	void testOneSecondFail() throws InterruptedException {
+		Thread.sleep(10_000);
+	}
+
+	@Test
+	@StrictTimeout(1)
+	void testOneSecondSuccess() throws InterruptedException {
+		Thread.sleep(800);
 	}
 }
