--- src/test/java/de/tum/in/testuser/subject/ArrayListUserProxy.java
@@ -1,10 +1,0 @@
-package de.tum.in.testuser.subject;
-
-import java.util.ArrayList;
-
-public class ArrayListUserProxy {
-	public static void useArrayList() {
-		ArrayList<String> x = new ArrayList<>();
-		x.add("abc");
-	}
-}
