+++ src/test/java/de/tum/in/testuser/subject/InputOutputPenguin.java
@@ -1,0 +1,26 @@
+package de.tum.in.testuser.subject;
+
+public class InputOutputPenguin extends MiniJava {
+
+	private InputOutputPenguin() {
+
+	}
+
+	public static void writeTwoLines() {
+		write("Pinguine sind die Besten!");
+		write("Nieder mit den Eisbären!");
+	}
+
+	public static void calculateSquare() {
+		int a = readInt("Zahl eingeben:");
+		write("Ausgabe:");
+		write(a * a);
+	}
+
+	public static void readTwoTimes() {
+		int a = readInt("Zahl eingeben:");
+		int b = readInt("Nächste");
+		write("Ausgabe:");
+		write(a * b);
+	}
+}
