--- src/test/java/de/tum/in/testutil/TestTest.java
@@ -1,20 +1,0 @@
-package de.tum.in.testutil;
-
-import static java.lang.annotation.ElementType.*;
-import static java.lang.annotation.RetentionPolicy.RUNTIME;
-
-import java.lang.annotation.Documented;
-import java.lang.annotation.Retention;
-import java.lang.annotation.Target;
-
-import org.junit.jupiter.api.Tag;
-import org.junit.jupiter.api.Test;
-
-@Test
-@Tag("test-test")
-@Documented
-@Retention(RUNTIME)
-@Target({ TYPE, METHOD, ANNOTATION_TYPE })
-public @interface TestTest {
-	// just a @Test @Tag combination
-}
